requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
gradio>=3.40.0
scikit-learn>=1.0.0